+++ fltk-derive/src/display.rs
@@ -209,7 +209,7 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
             fn buffer<'a>(&'a self) -> &'a mut TextBuffer {
                 unsafe {
                     let buffer = #get_buffer(self._inner);
-                    assert!(!buffer.is_null(), "Failed to get associated buffer!");
+                    assert!(!buffer.is_null());
                     let x = Box::from(TextBuffer::from_ptr(buffer));
                     &mut *Box::into_raw(x)
                 }
@@ -296,27 +296,43 @@ pub fn impl_display_trait(ast: &DeriveInput) -> TokenStream {
                 }
             }
 
-            fn move_right(&mut self) {
+            fn move_right(&mut self) -> Result<(), FltkError> {
                 unsafe {
-                    #move_right(self._inner);
+                    let x = #move_right(self._inner);
+                    if x == 0 {
+                        return Err(FltkError::Internal(FltkErrorKind::FailedOperation));
+                    }
+                    Ok(())
                 }
             }
 
-            fn move_left(&mut self){
+            fn move_left(&mut self) -> Result<(), FltkError> {
                 unsafe {
-                    #move_left(self._inner);
+                    let x = #move_left(self._inner);
+                    if x == 0 {
+                        return Err(FltkError::Internal(FltkErrorKind::FailedOperation));
+                    }
+                    Ok(())
                 }
             }
 
-            fn move_up(&mut self){
+            fn move_up(&mut self) -> Result<(), FltkError> {
                 unsafe {
-                    #move_up(self._inner);
+                    let x = #move_up(self._inner);
+                    if x == 0 {
+                        return Err(FltkError::Internal(FltkErrorKind::FailedOperation));
+                    }
+                    Ok(())
                 }
             }
 
-            fn move_down(&mut self){
+            fn move_down(&mut self) -> Result<(), FltkError> {
                 unsafe {
-                    #move_down(self._inner);
+                    let x = #move_down(self._inner);
+                    if x == 0 {
+                        return Err(FltkError::Internal(FltkErrorKind::FailedOperation));
+                    }
+                    Ok(())
                 }
             }