+++ src/menu.rs
@@ -64,6 +64,9 @@ impl MenuItem {
     
     /// Creates a popup menu at the specified coordinates and returns its choice
     pub fn popup(&mut self, x: i32, y: i32) -> Option<MenuItem> {
+        if self._inner.is_null() {
+            return None;
+        }
         unsafe {
             let item = Fl_Menu_Item_popup(self._inner, x, y);
             if item.is_null() {
@@ -78,18 +81,24 @@ impl MenuItem {
     }
     
     /// Returns the label of the menu item
-    pub fn label(&self) -> String {
+    pub fn label(&self) -> Option<String> {
+        if self._inner.is_null() {
+            return None;
+        }
         unsafe {
             let label_ptr = Fl_Menu_Item_label(self._inner);
-            assert!(!label_ptr.is_null(), "Failed to get menu item label!");
-            CStr::from_ptr(label_ptr as *mut raw::c_char)
+            if label_ptr.is_null() {
+                return None;
+            }
+            Some(CStr::from_ptr(label_ptr as *mut raw::c_char)
                 .to_string_lossy()
-                .to_string()
+                .to_string())
         }
     }
 
     /// Sets the label of the menu item
     pub fn set_label(&mut self, txt: &str) {
+        assert!(!self._inner.is_null());
         unsafe {
             let txt = CString::new(txt).unwrap();
             Fl_Menu_Item_set_label(self._inner, txt.as_ptr());
@@ -98,11 +107,13 @@ impl MenuItem {
 
     /// Returns the label type of the menu item
     pub fn label_type<T: WidgetType>(&self) -> T {
+        assert!(!self._inner.is_null());
         unsafe { T::from_i32(Fl_Menu_Item_label_type(self._inner)) }
     }
 
     /// Sets the label type of the menu item
     pub fn set_label_type<T: WidgetType>(&mut self, typ: T) {
+        assert!(!self._inner.is_null());
         unsafe {
             Fl_Menu_Item_set_label_type(self._inner, typ.to_int());
         }
@@ -110,36 +121,43 @@ impl MenuItem {
 
     /// Returns the label color of the menu item
     pub fn label_color(&self) -> Color {
+        assert!(!self._inner.is_null());
         unsafe { mem::transmute(Fl_Menu_Item_label_color(self._inner)) }
     }
 
     /// Sets the label color of the menu item
     pub fn set_label_color(&mut self, color: Color) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_set_label_color(self._inner, color as u32) }
     }
 
     /// Returns the label font of the menu item
     pub fn label_font(&self) -> Font {
+        assert!(!self._inner.is_null());
         unsafe { mem::transmute(Fl_Menu_Item_label_font(self._inner)) }
     }
 
     /// Sets the label font of the menu item
     pub fn set_label_font(&mut self, font: Font) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_set_label_font(self._inner, font as i32) }
     }
 
     /// Returns the label size of the menu item
     pub fn label_size(&self) -> u32 {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_label_size(self._inner) as u32 }
     }
 
     /// Sets the label size of the menu item
     pub fn set_label_size(&mut self, sz: u32) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_set_label_size(self._inner, sz as i32) }
     }
 
     /// Returns the value of the menu item
     pub fn value(&self) -> bool {
+        assert!(!self._inner.is_null());
         unsafe {
             match Fl_Menu_Item_value(self._inner) {
                 0 => false,
@@ -150,16 +168,19 @@ impl MenuItem {
 
     /// Sets the menu item
     pub fn set(&mut self) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_set(self._inner) }
     }
 
     /// Clears the menu item
     pub fn clear(&mut self) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_clear(self._inner) }
     }
 
     /// Returns whether the menu item is visible or not
     pub fn visible(&self) -> bool {
+        assert!(!self._inner.is_null());
         unsafe {
             match Fl_Menu_Item_visible(self._inner) {
                 0 => false,
@@ -170,6 +191,7 @@ impl MenuItem {
 
     /// Returns whether the menu item is active
     pub fn active(&mut self) -> bool {
+        assert!(!self._inner.is_null());
         unsafe {
             match Fl_Menu_Item_active(self._inner) {
                 0 => false,
@@ -180,21 +202,25 @@ impl MenuItem {
 
     /// Activates the menu item
     pub fn activate(&mut self) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_activate(self._inner) }
     }
 
     /// Deactivates the menu item
     pub fn deactivate(&mut self) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_deactivate(self._inner) }
     }
 
     /// Shows the menu item
     pub fn show(&mut self) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_show(self._inner) }
     }
 
     /// Hides the menu item
     pub fn hide(&mut self) {
+        assert!(!self._inner.is_null());
         unsafe { Fl_Menu_Item_hide(self._inner) }
     }
 }